Don't consider included options conflicting if they are the same value
In order to ensure users don't get confused where an option is
really being set, SAVE_CONFIG currently checks that include
files do not set options that are being saved in the main
config file's autosave section.

However, this means you can't set, say, the control mechanism
for a heater in an include file, because it is also
part of the autosave block.

This commit changes the conflict check to only consider it a
conflict if the included value differs from the one in the
autosave block (so it is okay if both set 'control:mpc' as an
example)

Signed-off-by: Daniel Berlin <dberlin@dberlin.org>