Support dynamic power for PTC heaters in MPC
This change adds basic support to MPC for heaters whose max power changes
as the temperature changes.
We require the voltage, temperature coefficient, and ambient temp at which
max wattage is spec'd (usually 23C).
With that, we compute the available power as the temperature changes,
instead of treating it as a constant.

Signed-off-by: Daniel Berlin <dberlin@dberlin.org>